Fully updated edges, distances and weights
-Changed edge parameters to doubles
-Added distanceCalc method to MapCanvas, that measures distance between to points using haversine formula
-Changed OSMWay constructor to take an integer speedlimit for the way.
-Parsing maxspeed from string to integer in loadOSM.

+++ app/src/main/java/bfst22/vector/Edge.java
@@ -4,7 +4,7 @@ import java.io.Serializable;
 
 public class Edge implements Serializable {
     private final long from, to;
-    private final float weight, distance;
+    private final double weight, distance;
     String name;
     //private final boolean forCar, forBike, forWalk;
 
@@ -13,8 +13,8 @@ public class Edge implements Serializable {
                 long from,
                 long to,
                 String name,
-                float weight,
-                float distance
+                double weight,
+                double distance
                 ){
 
         this.from = from;
@@ -31,10 +31,10 @@ public class Edge implements Serializable {
     public long getTo(){
         return to;
     }
-    public float getWeight(){
+    public double getWeight(){
         return weight;
     }
-    public float getDistance(){
+    public double getDistance(){
         return distance;
     }
